feat: support recording object identity references
V8 serialization format supports back-references to objects previously
occurring in the serialized data. In order to support this, the encoder
and decoder need to track references to objects they've seen, so that
the encoder can write back-references, and the decoder can resolve
back-references.

A side-effect of this is that data can have cycles (e.g. an object that
contains a reference to itself), which makes reading/writing immutable
data structures difficult. The encoder needs to record a reference to an
object before it has been populated with its entries, which makes using
regular immutable types like frozenset not possible.

For now, I'm opting to require regular mutable data types when
deserializing values.