feat: support multiple object mappers when encoding
The encoding process now supports layering more than one object mapper
in a chain, like middleware. This allows for object mappers to modify
serialization behaviour while delegating to the default behaviour for
other cases.

We use this to implement object back-references, as object references
need to check if any written object has already occurred, and if so
write a reference instead of serializing it again.

This will also allow for users to override behaviour without modifying
the default serialization definitions.