chore: make SimpleArrayProperties an ArrayProperties subclass
This is awkward because of Python prevents creating Protocol types that
subclasses MutableSequence, so we have to make ArrayProperties an ABC,
not a Protocol.

This sub SimpleArrayProperties class does not want to inherit from
MutableSequence ABC, so we have to hackaround with if TYPE_CHECKING to
trick the type checker into thinking it does extend ArrayProperties. 😒