feat: define SparseMutableSequence Protocol
We now define a Protocol for our sparse array, used to represent the
integer array properties of JavaScript objects.

It's somewhat tricky to define, because the stdlib typing module's
Sequence and MutableSequence are not Protocols, they're ABC classes. The
typing system does not allow creating classes that subclass both an ABC
and a Protocol, so you can't create a derived Sequence from the stdlib!

To work around this, I've copied the actual Sequence Protocol
definitions from typeshed into our own typing module, which allows us to
derive a new Protocol from Sequence.