feat: make element order ASCENDING by default
Before changing the 'reverse' kwarg to 'order' it was ascending by
default. I just made the mistake myself of forgetting to specify an
order when one was required (in __eq__), so I think it will be a source
of surprise to have different implementations return different orders by
default especially as lists seem like they should be ordered. This is
also consistent with Python's dict having a default predictable
(insertion order) ordering.

So now the UNORDERED order requires explicitly opting in.