feat: simplify JSObject types
JSObject is no longer generic in its key type. We had defined it as a
type var with base `str | int`, with the intention of letting users
constrain it to literals or just one or the other of int, str. In
practice it results in annoying, spurious type errors.

We now type the key as just str | int. We also overload methods that
accept collection that are generic in the str | int key type, as MyPy
does not seem to allow passing read-only interfaces with just one of
str, int to a method expecting to read a collection of str | int.