feat: support floats when normalising property keys
We now support floats as input as well as ints, because V8-serialized
JSObject data can store keys as floating point values. Handling these in
the same way as JavaScript requires some care, so by doing it here we
can remove the need for users or other parts of the API to know about
the differences.

The JavaScript behaviour depends on JavaScript's float to string
conversion, which is different to the default conversion in Python. We
replicate the behaviour for integer floating point values. We don't
attempt to match the string representation of arbitrary floats. Nobody
should be relying on a stringified float having an exact string
representation, unless they're doing something like hashing canonical
data, and in that case they are probably taking care of encoding data
themselves very carefully.