refactor: report raw and normalized key in KeyError
When throwing a KeyError from JSObject, we now use a specialised
exception that captures both the raw (provided) key, and the normalised
version that was actually looked up.

We do this now because we're going to support lookups with float values,
and the mapping of those to integers/strings may not be obvious if we
used the previous behaviour of reporting the raw (provided) key in the
KeyError. i.e. it would appear to an API user that the provided value
was looked up, not the actual string/int normalised value.