feat: support float in writes to JSObject
Because V8-serialized data can encode integer JSObject keys as floats,
the __setitem__/update/setdefault methods of JSObject now accept float
keys, which are normalised in the same way as str/int keys are.

I don't want to encourage use of floats as properties in general, so I'm
not adding float to the type of other methods, even though in practice
it will work, because all the methods normalise keys.

I'm not 100% sure adding support even to write methods is the right
thing to do, we could require the deserialiser to do the normalisation
itself...