feat: encode/decode JSArrayBuffer & Co
JSArrayBuffer now mimics regular JavaScript ArrayBuffer (providing
resize() etc) rather than just being a dumb container for a Python
buffer. The Resizable variant of JSArrayBuffer is now handled by
JSArrayBuffer.

Decoding ArrayBuffer types is improved by having the reader take a
function that creates an ArrayBuffer of the caller's choosing from the
decoded data. This allows the reader to record an object reference for
the created buffer. Previously the reader was creating a dumb struct of
data that the caller had to map to the final type themself.

When encoding, we now have a mechanism for de-duplicating values, which
allows for bytes/strings with distinct object identities but equivalent
values to be written once and referenced on subsequent occurrences. This
currently takes effect when we write bytes as ArrayBuffer. In the future
we could also de-dupe string values by writing them as heap objects.

JSArrayBuffer copies supplied data by default into a bytearray. This
results in intuitive default behaviour, as it results in a buffer
that's alway writeable, and doesn't depend on the source.

The copy_data=False option can be used to reference a source, as well as
readonly=True to make a source readonly via a readonly memoryview. This
also allows a JSArrayBuffer to adopt an existing bytearray without
copying it.