refactor!: remove flags from ArrayBuffer views
The flags for length tracking and buffer being resizable just duplicate
state that's determined by the backing buffer and the view's max length,
so flags just add a source of ambiguity/conflict. Instead we now derive
flags from the state when serializing and use flags to determine state
when deserializing.