feat: encode Python Exceptions as V8 stack traces
We now format Python exceptions in a format resembling the stack trace
format used by V8 for JavaScript errors. This results in V8 runtimes
like Node.js displaying Python errors in a readable way.

In contrast, if we use the default Python traceback format, Node.js
displays the exception in its REPL more like a JavaScript object, with
the stack string in brackets.