fix: handle ArrayBuffer[View] as error cause
The error cause encoding has the quirk that the byte after the cause
object is not a SerializationTag as it is in other situations. Instead
it's an Error End tag. This was tripping up the logic handling another
quirk of ArrayBufferView being able to occur immediately after another
object, because the Error End byte not being a serialization tag value
was triggering a validation error.

We now check for the tag type after an ArrayBuffer/ObjectReference in a
more lenient way.