feat: use a python-friendly == for JSMap & JSSet
Previously JSMap's __eq__ followed the same-value-zero equality of its
items. So two maps with seemingly identical but actually different
instances of the same object would not be equal. This behaviour is quite
surprising for people used to Python's normal equality.

Instead, the equality of keys internally is now separate from the
external equality of the overall JSMap. Two JSMap instances compare
equal when they contain items equal to each other in the same order.

This also has the nice property that it allows serialized and
deserialized data to remain equal, despite having been deep-copied, and
despite the JSMap being able to hold different instances of
equal-by-value keys.