feat: don't indent JS type repr by default
Although it's nice to view indented reprs of types, I'm not sure that it
should be the default. At least unless all of our types use the same
repr mechanism, not just the current collection types. Indented reprs
don't mix well when combined with repr of other types, unaware of the
indented layout.