test: Test serialization with real V8
We now run a round-trip serialization test using the actual V8
serialization logic. We use an external HTTP server running a javascript
program that deserializes a request and responds with the reserialized
value, along with the human-readable parsed interpretation of the data.

We use this in test_round_trip_with_v8 by generating example values with
our Hypothesis strategies, and then reflect them off the echo server,
checking that our starting value and deserialized result match.

This should fail if we've encoded values in a way that differs from V8's
understanding of the format. On the other hand, it can't detect problems
of mis-encoding values. For example, we could be encoding True as false
and false as True and not notice (without manually inspecting the V8
interpretation).