test: make theoretical V8 values optional in strategies
Test data strategies now allow theoretical V8 data to be disabled.
Theoretical values are things that are in principle possible to encode
in the V8 serialization format, but in practice are not received and
cause errors if sent. For example, objects which are references to
shared objects inside the V8 process, which V8 uses to share objects
between isolates.