refactor: rework and simplify JSArrayBufferView
The previous implementation was struggling to deal with the dynamic
behaviour of JavaScript TypedArrays when their backing buffer goes out
of range; and with validation of the initial view state.

The view types now persist the desired, initial view range in terms of
item units (not byte units). Using items avoids a class of errors with
specifying byte offsets/lengths that are not multiples of the item size.
We calculate the current byte offset/length on the fly, which makes it
natural to provide the JS behaviour of making the offset and length
become 0 when out of range.

We also separate JavaScript's initial view bounds-checking behaviour
from the normal constructor. The constructor allows views to be created
out of range. A separate factory function creates views while enforcing
JavaScript's requirement that the view starts in-range (even though it
allows views to later become out-of-range).