feat: report supported features & versions from echoserver
The echoserver now responds to GET / with metadata containing the
node/deno/v8 versions it's running, as well as the SerializationFeatures
it supports. The server detects supported serialization features
automatically by attempting to serialize data that uses the features.

This should allow us to dynamically update and add new versions of
runtimes, without hard-coding versions and features.