feat: allow users to change repr settings of JS types
We now provide a context manager js_repr_settings() which changes the
settings of the reprlib.Repr instance we use to generate the repr of
various JS types.

Also, repr tests now use insta snapshots rather than doctests, as
they're easier to update/manage, and we already use insta for other
snapshot tests, so it makes sense to unify.