chore: fix inherited descriptor breaking dataclass in py3.9
Inheriting a @dataclass class from Protocol classes that contain
read-only descriptors prevents the derived class from setting a value
for a property named the same as the inherited descriptor.

For example, JSArrayBuffer was inheriting the AnyArrayBuffer Protocol
explicitly, to document it and enforce type compatibility. The
protocol's fields get picked up by @dataclass, which prevents setting
max_byte_length in init.

We also had this with JSErrorData and solved it by creating the
dataclass as a standalone class and inheriting it with the protocol
separately. In this case I'm taking the approach of using a type
assertion in a TYPE_CHECKING block to ensure that the protocol is
satisfied.

I've created a test case to describe this behaviour and some solutions.
The best workaround depends a bit on the situation, but either declaring
@property in protocols only if TYPE_CHECKING works if @runtime_checkable
isn't required, otherwise overriding the descriptor with ...
if not TYPE_CHECKING works for @runtime_checkable protocols.