chore: handle IntFlag ~ behaviour in py3.9
In py12 negating an empty IntFlag gives a flag with all members set (the
max value). In 3.9 it negates the 0 int value, giving -1. We now use
sum() to get a flag with all members set. I don't feel the need to
define a constant for this, as in practice setting every flag is
invalid.