refactor: include __doc__ in both _typing types
The _typing module defines each Protocol twice, once for runtime and
once for type-checking needs. Unfortunately generating docs from
docstrings adds another spanner in the works.

We can either use dynamic extraction of __doc__ by executing the module
and pulling out the actual __doc__ attrs (this uses the non
TYPE_CHECKING branch), or use static analysis to read docstrings from
the source (which seems to use the last-defined version in the source
file).

I'm taking the approach of making both runtime and typing branches as
close to a copy of each other as possible. The only differences should
be not inheriting the collections.abc types at runtime, and not using
@property at runtime.