test: fix flaky fail in JSObjectComparisonMachine
Hypothesis found a failure scenario in the stateful property test using
property '00'. The test's model was incorrectly treating it as an
integer property, when in fact it should be (and is) treated as a string
property. This caused the test to fail incorrectly.

The test now uses the well-tested normalise_property_key() function
instead of int() to distinguish integer from non-integer properties.