chore: fix mypy failing to infer type of read_js_array_buffer
Mypy was inferring the return type of read_js_array_buffer as ABC
instead of the union of the 3 arguments. (ABC is a base class of all 3
arguments passed at at the call site.) Separating the 3 arguments into
3 separate type vars seems to help it keep the 3 types distinct rather
than resolve the common ancestor.

Additionally, mypy also seems to need an explicit subscript to declare
the generic type of `JSArrayBuffer[ReadableBinary]`, even though
`ReadableBinary` *is* the default type of the typevar, so this shouldn't
be necessary.