chore: add vscode devcontainer config
This is the devcontainer config I use to work on the project. It has some
config to allow the docker compose config to interoperate with the
devcontainer container, by putting the devcontainer on the same docker
network as the compose config, and mounting the devcontainer's workspace
volume into the `integration_test` service.

It also also supports GitHub codespaces, which are quite similar to
devcontainers.